extract MajorMinorVersion
